requirements.txt
ffmpeg-python
scipy
moviepy
ipywidgets